python_backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.4.2
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

